+++ cloud_client.go
@@ -34,7 +34,7 @@ func NewClient(config Config) *CloudClient {
 
 	timeout := config.TimeoutSecond
 	if timeout <= 0 {
-		timeout = 30
+		timeout = 10
 	}
 
 	client.HttpClient = &http.Client{
@@ -100,9 +100,9 @@ func (cloudClient *CloudClient) Request(method string, requestPath string, param
 
 	// set header
 	if auth == NONE {
-		Headers(request, "", "", "")
+		Headers(request, "", "", "", config.Headers)
 	} else if auth == KEYED {
-		Headers(request, config.ApiKey, "", "")
+		Headers(request, config.ApiKey, "", "", config.Headers)
 	} else if auth == SIGNED {
 		timestamp := UTCTime()
 		sign, err := HmacSha256Base64Signer(
@@ -110,7 +110,7 @@ func (cloudClient *CloudClient) Request(method string, requestPath string, param
 		if err != nil {
 			return response, err
 		}
-		Headers(request, config.ApiKey, timestamp, sign)
+		Headers(request, config.ApiKey, timestamp, sign, config.Headers)
 	}
 
 	if config.IsPrint {
@@ -136,11 +136,12 @@ func (cloudClient *CloudClient) Request(method string, requestPath string, param
 		return response, err
 	}
 
-	cloudResponse.httpStatus = response.StatusCode
-	cloudResponse.response = string(body)
-	cloudResponse.limit.limit = StringToInt(response.Header.Get("X-BM-RateLimit-Limit"))
-	cloudResponse.limit.remaining = StringToInt(response.Header.Get("X-BM-RateLimit-Remaining"))
-	cloudResponse.limit.reset = StringToInt(response.Header.Get("X-BM-RateLimit-Reset"))
+	cloudResponse.HttpStatus = response.StatusCode
+	cloudResponse.Response = string(body)
+	cloudResponse.Limit.Limit = StringToInt(response.Header.Get("X-BM-RateLimit-Limit"))
+	cloudResponse.Limit.Remaining = StringToInt(response.Header.Get("X-BM-RateLimit-Remaining"))
+	cloudResponse.Limit.Reset = StringToInt(response.Header.Get("X-BM-RateLimit-Reset"))
+	cloudResponse.Limit.Mode = response.Header.Get("X-BM-RateLimit-Mode")
 
 	return response, nil
 }